+++ tower-fallthrough-filter/Cargo.toml
@@ -15,16 +15,17 @@ categories = ["asynchronous", "network-programming"]
 [dependencies]
 futures = "0.3.30"
 tower = "0.4.13"
+pin-project = { version = "1.1.4", optional = true }
 
 [dev-dependencies]
 axum = "0.7.4"
 axum-test = "14.3.1"
-pin-project = "1.1.4"
 tokio = { version = "1.36.0", features = ["full"] }
 
 [features]
 default = []
-async = []
+futures = [ "dep:pin-project" ]
+async = [ "futures" ]
 
 [[example]]
 name = "axum-render-layer-async"
